test: fix compilation errors + validate Week 9 architecture (258/279 pass)
- Fixed missing Deserialize derives in enhanced_list_files.rs
- Fixed ToolOutput pattern matching in skills/tool.rs
- Fixed type mismatches in performance_benchmarks.rs (Vec<u128> -> Vec<f64>)
- Validated: 17/17 Week 7-8 integration tests pass (hybrid architecture working)
- Validated: 258/279 total tests pass (92.5% pass rate)
- Created: ai/TEST_RESULTS_2025-10-30.md (comprehensive test analysis)
- Updated: ai/VALIDATION_ASSESSMENT.md with actual test data

Key findings:
✅ Hybrid architecture IS INTEGRATED (event bus, mode enforcement, model routing all working)
✅ Core functionality validated (92.5% tests pass)
✅ Zero compilation errors
⚠️ 21 edge case failures (minor, non-critical)
❌ End-to-end benefits unproven (60% reduction, 90% speedup - need full agent execution)

Next: Performance benchmarks for measurable components

+++ src/testing/performance_benchmarks.rs
@@ -132,8 +132,8 @@ impl PerformanceBenchmarks {
     async fn benchmark_response_times(&mut self) -> Result<()> {
         println!("      ⏱️  Benchmarking response times...");
 
-        let mut p50_times: Vec<u128> = Vec::new();
-        let mut p95_times: Vec<u128> = Vec::new();
+        let mut p50_times: Vec<f64> = Vec::new();
+        let mut p95_times: Vec<f64> = Vec::new();
 
         // Simulate various query types
         for i in 0..self.config.performance_iterations {
